refactor: pricePerNight 변수 타입 수정

+++ src/main/java/guesthouse/guestHousePost/domain/model/Room.java
@@ -39,5 +39,5 @@ public class Room {
     private LocalTime checkOutTime;
 
     @Column(nullable = false)
-    private Long pricePerNight;
+    private Integer pricePerNight;
 }
